feat: Implemented coral intake command
When R1 is held, the intake spins forward and automatically slows down when the coral is detected in the intake. When R2 is held, the intake reverses at a constant speed. When one button is being held and the other is pressed, the intake mode does not change until the original button is released. Also fixed errors with SVA constant orders.