Logging for Arm Command + Neutral Positions
I'm not there on Wednesday so I'm doing some of the work I wanted to do then.
Important: None of this code is functional unless the team wants it to be functional. Most of it (except logging) is put behind a boolean that is always set to false so the code I added won't run unless you consciously set it to true.

CHANGES:
1.Added logging to ArmCommand. The target pivot angle, elevator length, and wrist angle are now always logged in AdvantageScope.
2. Added neutral positions. After the robot does some specific actions with the intake, the robot now goes into a neutral position that changes based on what is in the intake. Comments in the ArmCommand file. This code is locked behind a boolean so it won't run unless you guys absolutely want it to be tested.
3. Added a hasCoralBurger and hasCoralHotDog. They return true based on whether our intake has coral "hot dog" style or "burger" style. Specifics in CoralIntake file.
4. Added setting the arm command to neutral after the coral intake and algae intake is run in ControlMap. This won't do anything unless those positions are enabled so it won't mess with other more important code we want to test.