Added units to pivot
Changed doubles to java units for the pivot, renamed variables for clarity (e.g. angle to pivotAngle), changed pivotAppliedVolts to motorAppliedVoltage for consistency, and fixed elevator use of pivot methods to work with the new units.

TODO: We should probably add a custom units class in utils where we can define units such as jerk and the feedforward units so they aren't so long